+++ __tests__/GameController.spec.ts
@@ -1,141 +1,8 @@
 import GameController, { isGameState } from '@src/controller/GameController';
-import { GameState } from '@src/model/GameState';
-import { Event } from '@src/model/Events';
-
-const simpleBinaryChoiceNarrative: Event[] = [
-    {
-        id: 'start',
-        name: 'Start of narrative',
-        description: 'A fancy text',
-        canRun: (gameState: GameState) => {
-            return gameState.turnNumber === 1;
-        },
-        responses: [
-            {
-                id: 'start.goLeft',
-                eventId: 'start',
-                name: 'Go left',
-                label: ['Go left'],
-                isApplicable: (gameState: GameState) => true,
-                onSelect: (gameState: GameState) => ({
-                    updatedIndicators: {
-                        ...gameState.indicators,
-                        reputation: [
-                            {
-                                id: 'left',
-                                name: 'Left',
-                                icon: 'string',
-                                description: 'string'
-                            }
-                        ]
-                    },
-                    feedback: {
-                        toResponse: 'left',
-                        fromPublic: 'left',
-                        fromBusiness: 'left',
-                        fromHealthcare: 'left'
-                    }
-                })
-            },
-            {
-                id: 'start.goRight',
-                eventId: 'start',
-                name: 'Go right',
-                label: ['Go right'],
-                isApplicable: (gameState: GameState) => true,
-                onSelect: (gameState: GameState) => ({
-                    updatedIndicators: {
-                        ...gameState.indicators,
-                        reputation: [
-                            {
-                                id: 'right',
-                                name: 'Right',
-                                icon: 'string',
-                                description: 'string'
-                            }
-                        ]
-                    },
-                    feedback: {
-                        toResponse: 'right',
-                        fromPublic: 'right',
-                        fromBusiness: 'right',
-                        fromHealthcare: 'right'
-                    }
-                })
-            }
-        ]
-    },
-    {
-        id: 'left',
-        name: 'Left branch',
-        description: 'A fancy text',
-        canRun: (gameState: GameState) => {
-            if (gameState.responseHistory.length < 1) {
-                return false;
-            } else {
-                const wentLeft = gameState.responseHistory[gameState.responseHistory.length - 1].responses.find(
-                    (resp) => resp.response.id === 'start.goLeft'
-                );
-                return gameState.turnNumber === 2 && wentLeft !== undefined;
-            }
-        },
-        responses: [
-            {
-                id: 'left.end',
-                eventId: 'left',
-                name: 'Finish',
-                label: ['Narrative ended on the left branch'],
-                isApplicable: (gameState: GameState) => true,
-                onSelect: (gameState: GameState) => ({
-                    updatedIndicators: gameState.indicators,
-                    feedback: {
-                        toResponse: 'done left',
-                        fromPublic: 'done left',
-                        fromBusiness: 'done left',
-                        fromHealthcare: 'done left'
-                    }
-                })
-            }
-        ]
-    },
-    {
-        id: 'right',
-        name: 'Right branch',
-        description: 'A fancy text',
-        canRun: (gameState: GameState) => {
-            if (gameState.responseHistory.length < 1) {
-                return false;
-            } else {
-                const wentRight = gameState.responseHistory[gameState.responseHistory.length - 1].responses.find(
-                    (resp) => resp.response.id === 'start.goRight'
-                );
-                return gameState.turnNumber === 2 && wentRight !== undefined;
-            }
-        },
-        responses: [
-            {
-                id: 'right.end',
-                eventId: 'right',
-                name: 'Finish',
-                label: ['Narrative ended on the right branch'],
-                isApplicable: (gameState: GameState) => true,
-                onSelect: (gameState: GameState) => ({
-                    updatedIndicators: gameState.indicators,
-                    feedback: {
-                        toResponse: 'done right',
-                        fromPublic: 'done right',
-                        fromBusiness: 'done right',
-                        fromHealthcare: 'done right'
-                    }
-                })
-            }
-        ]
-    }
-];
+import { simpleBinaryChoiceNarrative, testScenario } from './Fixtures';
 
 describe('The operation of the game controller', () => {
-    const narrative = simpleBinaryChoiceNarrative;
-    const gameController = new GameController(narrative);
+    const gameController = new GameController(testScenario, simpleBinaryChoiceNarrative);
 
     it('Is able to start', () => {
         expect(gameController.canProceedToNextTurn()).toBe(true);
