+++ src/scenarios/US_runup.ts
@@ -1,0 +1,405 @@
+export const usHistory = [
+    {
+        days: -31,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -30,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -29,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -28,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -27,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -26,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -25,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -24,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -23,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -22,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -21,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -20,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -19,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -18,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -17,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -16,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -15,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -14,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -13,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -12,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -11,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -10,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -9,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -8,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -7,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -6,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -5,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -4,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -3,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -2,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    },
+    {
+        days: -1,
+        numDead: 2504,
+        numInfected: 250443,
+        totalPopulation: 400000000,
+        hospitalCapacity: 250000,
+        r: 1.09,
+        importedCasesPerDay: 0.1,
+        deathCosts: 25043932258,
+        economicCosts: 10576809135,
+        medicalCosts: 1252215000,
+        totalCost: 36872956393
+    }
+];
